API call working, adding marker with delay working. max items to add is 50, need to check earlier items maybe

+++ app/src/main/java/com/example/sensorbasedmobileproject/data/NominatimItem.kt
@@ -11,7 +11,7 @@ data class NominatimItem(
     val place_id: Int,
     val licence: String,
     val osm_type: String,
-    val osm_id: Int,
+    val osm_id: Long,
     // val boundingbox: List<Double>,
     val lat: Double,
     val lon: Double,
